fixed codeclimate
