+++ packages/simcore-sdk/src/simcore_sdk/node_ports_common/constants.py
@@ -2,5 +2,6 @@ from typing import Final
 
 from models_library.api_schemas_storage import LocationID
 
-SIMCORE_LOCATION: Final[LocationID] = 0
+CHUNK_SIZE: Final[int] = 16 * 1024 * 1024
 MINUTE: Final[int] = 60
+SIMCORE_LOCATION: Final[LocationID] = 0
